# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Peter Dashboard API
  description: |
    Backend API for Peter dashboard platform.
    Handles dashboard management, BigQuery query execution, authentication, and lineage tracking.
  version: 0.1.0
  contact:
    name: Peter Team
    email: team@peter.com
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api-staging.peter.com
    description: Staging environment
  - url: https://api.peter.com
    description: Production environment
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: peter_session
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    DateTime:
      type: string
      format: date-time
      example: "2025-10-30T10:15:30Z"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid dashboard YAML"
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
          example: "abc123"
        remediation:
          type: string
          example: "Check YAML syntax and try again"

    # Auth models
    UserInfo:
      type: object
      required:
        - id
        - email
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          nullable: true
          example: "John Doe"
        created_at:
          $ref: "#/components/schemas/DateTime"
        last_login:
          $ref: "#/components/schemas/DateTime"
          nullable: true

    # Dashboard models
...
paths:
...
  /api/v1/auth/login:
    get:
      tags: [auth]
      summary: Initiate OAuth login
      operationId: initiateLogin
      responses:
        "302":
          description: Redirect to Google OAuth
        "500":
          description: OAuth configuration error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/callback:
    get:
      tags: [auth]
      summary: OAuth callback
      operationId: handleCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to app with session cookie
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out successfully
        "401":
          description: Not authenticated
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Dashboard endpoints
...